pyproject.toml
[project]
name = "quickbite"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.16.4",
    "argon2-cffi>=25.1.0",
    "asyncpg>=0.30.0",
    "celery>=5.5.3",
    "fastapi[all]>=0.116.1",
    "flower>=2.0.1",
    "geoalchemy>=0.7.2",
    "geoalchemy2>=0.18.0",
    "passlib>=1.7.4",
    "pre-commit>=4.2.0",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.1.1",
    "python-jose[cryptography]>=3.5.0",
    "redis>=6.2.0",
    "ruff>=0.12.7",
    "shapely>=2.1.1",
    "sqlalchemy[asyncio]>=2.0.41",
]

# Enable rules
lint.select = ["E", "F", "I", "N", "UP", "B", "C90", "RUF"]

lint.ignore = [
    "E501",     # line too long
    "RUF001",   # ambiguous unicode character in code
    "RUF002",   # ambiguous unicode character in docstring\
    "RUF012",
]  

# Paths to exclude
exclude = ["migrations", "__pycache__", ".venv"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "lf"

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra --tb=short --strict-markers"
testpaths = ["tests"]
